Updated tests model

+++ models/test.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Double
 from sqlalchemy.orm import relationship
 from db.db import Base
 
@@ -6,10 +6,10 @@ class Test(Base):
     __tablename__ = "tests"
     
     id = Column(Integer, primary_key=True, index=True)
-    reading_score = Column(Integer, nullable=False)
-    writing_score = Column(Integer, nullable=False)
-    listening_score = Column(Integer, nullable=False)
-    speaking_score = Column(Integer, nullable=False)
+    reading_score = Column(Double, nullable=False)
+    writing_score = Column(Double, nullable=False)
+    listening_score = Column(Double, nullable=False)
+    speaking_score = Column(Double, nullable=False)
     
     student_id = Column(Integer, ForeignKey("students.id"), nullable=False)
